fix: reset all e reset group com 10 reles

+++ src/config/ConfigManager.h
@@ -18,7 +18,7 @@ public:
     
     void loadHardwareState(DeviceState& state);
     void saveRelayState(uint8_t index, bool state);
-    void saveRelayMask(uint8_t mask);
+    void saveRelayMask(uint16_t mask);
     void saveDACState(uint8_t channel, float voltage);
     void resetHardwareState();
